Creating data base for pages and gallery
Today i have learnt how to insert data from my webpage into a database through sql. I learnt how to create variables that link to the database so that i can have a products page that functions through the use of a database. This is helpful so that my code is not hardwired and can have products changed added, and found more easily.