Changed gallery titles to match category
I used php to say if a category is set, echo a different statement for each different category. I did this in the <h2> tags for my gallery page.
i also have been in the process of removing infomation from the div class "gallery" and inserting it into the products table in the database