Made more admin accessible pages that show database
I created a admin folder for pages that can only be accessed by the admin.  In this folder i created an invoice.php displaying a table of the invoices in the database. I also created pages.php which allows the admin to edit the website when logged in as admin